プロジェクト周りのコードを整理
