+++ src/components/Sing/SequencerNote.vue
@@ -66,7 +66,7 @@
 <script setup lang="ts">
 import { computed, ref } from "vue";
 import { useStore } from "@/store";
-import { Note } from "@/store/type";
+import { Note } from "@/domain/project/type";
 import {
   getKeyBaseHeight,
   tickToBaseX,
