+++ src/components/Sing/SequencerShadowNote.vue
@@ -15,7 +15,7 @@
 <script setup lang="ts">
 import { computed } from "vue";
 import { useStore } from "@/store";
-import { Note } from "@/store/type";
+import { Note } from "@/domain/project/type";
 import {
   getKeyBaseHeight,
   tickToBaseX,
