+++ src/domain/project/type.ts
@@ -1,0 +1,25 @@
+import { z } from "zod";
+
+import {
+  noteSchema,
+  phonemeTimingEditSchema,
+  singerSchema,
+  tempoSchema,
+  timeSignatureSchema,
+  trackSchema,
+} from "@/domain/project/schema";
+import { NoteId } from "@/type/preload";
+
+export type Tempo = z.infer<typeof tempoSchema>;
+
+export type TimeSignature = z.infer<typeof timeSignatureSchema>;
+
+export type Note = z.infer<typeof noteSchema>;
+
+export type Singer = z.infer<typeof singerSchema>;
+
+export type Track = z.infer<typeof trackSchema>;
+
+export type PhonemeTimingEdit = z.infer<typeof phonemeTimingEditSchema>;
+
+export type PhonemeTimingEditData = Map<NoteId, PhonemeTimingEdit[]>;
