+++ src/sing/sequencerStateMachine/states/addNoteState.ts
@@ -8,7 +8,7 @@ import {
   SequencerStateDefinitions,
 } from "@/sing/sequencerStateMachine/common";
 import { NoteId, TrackId } from "@/type/preload";
-import { Note } from "@/store/type";
+import { Note } from "@/domain/project/type";
 import {
   getButton,
   getDoremiFromNoteNumber,
