+++ src/sing/sequencerStateMachine/states/resizeNoteLeftState.ts
@@ -1,7 +1,7 @@
 import { getOrThrow } from "@/helpers/mapHelper";
 import { State, SetNextState } from "@/sing/stateMachine";
 import { getButton, PREVIEW_SOUND_DURATION } from "@/sing/viewHelper";
-import { Note } from "@/store/type";
+import { Note } from "@/domain/project/type";
 import { TrackId, NoteId } from "@/type/preload";
 import {
   Context,
