+++ src/sing/utaformatixProject/toVoicevox.ts
@@ -3,7 +3,7 @@ import { VoicevoxScore } from "./common";
 import { DEFAULT_TPQN, createDefaultTrack } from "@/sing/domain";
 import { getDoremiFromNoteNumber } from "@/sing/viewHelper";
 import { NoteId } from "@/type/preload";
-import { Note, Tempo, TimeSignature, Track } from "@/store/type";
+import { Note, Tempo, TimeSignature, Track } from "@/domain/project/type";
 import { uuid4 } from "@/helpers/random";
 
 /** UtaformatixのプロジェクトをVoicevoxの楽譜データに変換する */
