+++ src/store/project/index.ts
@@ -11,16 +11,12 @@ import {
   AudioItem,
   ProjectStoreState,
   ProjectStoreTypes,
-  Track,
 } from "@/store/type";
 import { TrackId } from "@/type/preload";
 import path from "@/helpers/path";
 import { getValueOrThrow, ResultError } from "@/type/result";
-import { LatestProjectType } from "@/domain/project/schema";
-import {
-  migrateProjectFileObject,
-  ProjectFileFormatError,
-} from "@/domain/project";
+import { LatestProjectType } from "@/infrastructures/projectFile/type";
+import { ProjectFileFormatError } from "@/infrastructures/projectFile/type";
 import {
   createDefaultTempo,
   createDefaultTimeSignature,
@@ -36,6 +32,8 @@ import {
 } from "@/components/Dialog/Dialog";
 import { uuid4 } from "@/helpers/random";
 import { recordToMap } from "@/sing/utility";
+import { Track } from "@/domain/project/type";
+import { migrateProjectFileObject } from "@/infrastructures/projectFile/migration";
 
 export const projectStoreState: ProjectStoreState = {
   savedLastCommandIds: { talk: null, song: null },
