+++ src/store/type.ts
@@ -63,19 +63,19 @@ import {
   ConfirmDialogOptions,
   WarningDialogOptions,
 } from "@/components/Dialog/Dialog";
-import {
-  LatestProjectType,
-  noteSchema,
-  phonemeTimingEditSchema,
-  singerSchema,
-  tempoSchema,
-  timeSignatureSchema,
-} from "@/domain/project/schema";
 import { HotkeySettingType } from "@/domain/hotkeyAction";
 import {
   MultiFileProjectFormat,
   SingleFileProjectFormat,
 } from "@/sing/utaformatixProject/utils";
+import {
+  Note,
+  Singer,
+  Tempo,
+  TimeSignature,
+  Track,
+} from "@/domain/project/type";
+import { LatestProjectType } from "@/infrastructures/projectFile/type";
 
 /**
  * エディタ用のAudioQuery
@@ -749,34 +749,6 @@ export type AudioPlayerStoreTypes = {
 /*
  * Singing Store Types
  */
-
-// schemaはプロジェクトファイル用
-export type Tempo = z.infer<typeof tempoSchema>;
-
-export type TimeSignature = z.infer<typeof timeSignatureSchema>;
-
-export type Note = z.infer<typeof noteSchema>;
-
-export type Singer = z.infer<typeof singerSchema>;
-
-export type Track = {
-  name: string;
-  singer?: Singer;
-  keyRangeAdjustment: number; // 音域調整量
-  volumeRangeAdjustment: number; // 声量調整量
-  notes: Note[];
-  pitchEditData: number[]; // 値の単位はHzで、データが無いところはVALUE_INDICATING_NO_DATAの値
-  phonemeTimingEditData: Map<NoteId, PhonemeTimingEdit[]>;
-
-  solo: boolean;
-  mute: boolean;
-  gain: number;
-  pan: number;
-};
-
-export type PhonemeTimingEdit = z.infer<typeof phonemeTimingEditSchema>;
-
-export type PhonemeTimingEditData = Map<NoteId, PhonemeTimingEdit[]>;
 
 export type PhraseState =
   | "SINGER_IS_NOT_SET"
