+++ tests/unit/domain/project.spec.ts
@@ -1,6 +1,6 @@
 import fs from "fs";
 import { beforeEach, describe, expect, test } from "vitest";
-import { migrateProjectFileObject } from "@/domain/project";
+import { migrateProjectFileObject } from "@/infrastructures/projectFile/migration";
 import { EngineId, SpeakerId, StyleId } from "@/type/preload";
 import { resetMockMode } from "@/helpers/random";
 import path from "@/helpers/path";
